make IAM policy structs reusable
